Step 1: Refactored the `User` class
Made it so that the User class is now the Account class. This is to
prevent confusion between the User class and the Spring Security User
class. Also split up the responsibilities of the User class into the
Account class and the Profile class. The Account class is now responsible
for the authentication of the user and the Profile class is responsible
for the user's profile information.

Next up will be to change all other classes using the User class to use
the Account class instead.