Completion of the Account classes
Completed the migration and refactoring of the User class to Account,
and the relevant service and repository classes. As such the User class
has been removed in favor of the Account class.

Updated the fields in the Itinerary classes to more accurately reflect
the data that will be stored in the database and to be more realistic
in a sense.

Authentication is not fully complete for the ItineraryController, but
you can now create, read and delete an itinerary for a user.
You can't update entries but will be implemented soon.

Still no tests, but will be implemented soon.