Support for authorization with Itineraries.
Added a PermissionEvaluator to check if a user has permission to an
itinerary ('read', 'write', 'delete'). This is untested and will be
tested in the future.

Removed the possibility to change author of an itinerary. But will
probably be added again in the future with more robust methods.

Renamed Constants to Util and added a method to get the current date in
a SQL supported format.