Packaging Refactor Part 1
First part of the packaging refactor. This commit is a bit messy and
uncomplete, as such many imports will be broken and other code.
This commit is to split the refactored commits.

Removed the collaborators from Itinerary and Accounts as to speed up CRUD
operations. As collaborators aren't needed. (This may change in the future)
As such I've updated the relevant classes that used the collaborators
collection.

Added @Column annotations to the Account classes. This is to make it more
clear what the column can be null or not. (Itinerary will be done in the
future)

Added the prefix 'ROLE_' to all roles in the database. This is to
make it more standardised with the Spring Security framework.