Packaging Refactor Part 2
Second part of the packaging refactor. This commit is still a bit messy,
but it's a lot better than the last one. Next one is the last part.

Renamed the mappings of the ItineraryController to be more readable. I
also added a public/unsecured controller for the public api (such as
getting all public itineraries of a user). As such I've updated the
permissions to include a public endpoint ('api/v1/public/')

Also the permissions is now using a switch statement with an enum
instead of a bunch of if statements.