Login Part 1 & Front Page Part 3
Added a custom button util file which will create buttons and links with my style every time. Updated the image slideshow to be more scalable. Need to update the responsiveness of the front page again.

Added some documentation.

Added a Login Page and a request to login, changes were made to the backend to support requests from the site (locally for now). A successful login attempt will add the token as a cookie.