Refactor of the Account Entity
A big refactor the the `Account` entity, because the general lack of readability and consistency of the classes before. Changed from `Date` to `Timestamp` for more accurate timings. As such this major refactor also impacted other classes such as the `Service` classes and `Controller` classes. Removed the `Location` class and moved the `country` field to `Details`

Updated User Tables Purpose:
- Account: Private Information.
- Profile: Public and Optional Private Information.
- Details: Optional Public Information.
- Verification: Details regarding user verification.

Added the Quartz Scheduler to automatically update the `Verification` table.

Updated the `Response` and `Request` to streamline logging in and registering a new user.

Added a new constants class for error messages.